+++ Src/Couchbase/CouchbaseBucket.cs
@@ -623,7 +623,7 @@ namespace Couchbase
         /// <param name="cas">The CAS (Check and Set) value for optimistic concurrency.</param>
         /// <param name="expiration">The time-to-live (ttl) for the key in seconds.</param>
         /// <returns>An object implementing the <see cref="IOperationResult{T}"/>interface.</returns>
-        public IOperationResult<T> Replace<T>(string key, T value, uint cas, uint expiration)
+        public IOperationResult<T> Replace<T>(string key, T value, ulong cas, uint expiration)
         {
             var operation = new Replace<T>(key, value, null, _converter, _transcoder)
             {
